Re-organize file naming schemes
