+++ frontend/src/components/app.js
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact';
 import { Router } from 'preact-router';
 
-import Header from './header';
+import Navbar from './Navbar';
 
 // Code-splitting is automated for routes
 import Home from '../routes/home';
@@ -19,7 +19,7 @@ export default class App extends Component {
   render() {
     return (
       <div id="app">
-        <Header />
+        <Navbar />
         <Router onChange={this.handleRoute}>
           <Home path="/" />
           <Profile path="/profile/" user="me" />
