+++ frontend/src/index.js
@@ -1,4 +1,4 @@
 import './style';
-import App from './components/app';
+import App from './components/App';
 
 export default App;
